Moved User struct to model folder
