+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod db;
+pub mod model;
 pub mod router;
 pub mod server;
 pub mod settings;
