+++ src/model/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod user;
+
+pub use user::User;
